+++ src/uPhp/u.php
@@ -30,6 +30,7 @@ final class u
         static::$classMap = include(UPHP_PATH . '/classes.php');
 
         foreach (array_keys(static::$classMap) as $classNamespace) {
+            print_r($name);
             if (method_exists($classNamespace, $name)) {
                 return call_user_func_array([$classNamespace, $name], $arguments);
             }
@@ -75,7 +76,7 @@ final class u
      */
     public static function version()
     {
-        return '0.1.0';
+        return '0.1.2';
     }
 
     /**