+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 Change Log
 ==========
 
+2.0.0
+-----
+
+* Reformat utils
+
+
+0.1.2
+-----
+* Add function to array: `Arrays::exists($key, $array)` 
+* Add function to array: `Arrays::save($key, &$array, $value, $replace = true)` 
+* Add function to array: `Arrays::delete($key, &$array)` 
+* Add function to array: `Arrays::get($key, $array, $default = null, $ignoreString = true)` 
+
 0.1.0
 -----